+++ CHANGELOG.md
@@ -4,7 +4,9 @@ Change Log
 Version 1.3.0 *(2014-XX-XX)*
 ----------------------------
 * LIB: Added a CardGridArrayMultiChoiceAdapter [(doc)](https://github.com/gabrielemariotti/cardslib/tree/master/doc/CARDGRID.md#using-a-cardgrid-in-multichoicemode)
+* LIB: Added a CardArrayMultiChoiceAdapter [(doc)](https://github.com/gabrielemariotti/cardslib/tree/master/doc/CARDLIST.md#using-a-cardlist-in-multichoicemode)
 * DEMO: Added an example with CardGridArrayMultiChoiceAdapter
+* DEMO: Added an example with CardArrayMultiChoiceAdapter
 
 
 Version 1.2.0 *(2014-01-07)*